refactor(counters): move value-write action bodies to counter-actions
The route still carried the set/addEntry/deleteEntry logic, keeping its
overall complexity above main's. Move those three bodies into
lib/server/counter-actions.ts (runSet/runAddEntry/runDeleteEntry) so the
actions map is one-line wiring — the route is now simpler than before the
feature. No behaviour change.
